Initialized KeyGen class. Moved key_randomizer and split_keys methods to KeyGen class.
